Fix offsets using ordinals. - Port over loading MPQs and the stub function associated. - Modules are loaded as soon as the first D2 variable is requested.
