+++ src/D2Offset.cpp
@@ -26,13 +26,37 @@
 #include "D2Offset.h"
 #include "DLLmain.h"
 
+DLLBaseStrc gptDllFiles[] = {
+        { L"Binkw32.dll", nullptr },
+        { L"BnClient.dll", nullptr },
+        { L"D2Client.dll", nullptr },
+        { L"D2CMP.dll", nullptr },
+        { L"D2Common.dll", nullptr },
+        { L"D2DDraw.dll", nullptr },
+        { L"D2Direct3D.dll", nullptr },
+        { L"D2Game.dll", nullptr },
+        { L"D2Gdi.dll", nullptr },
+        { L"D2Gfx.dll", nullptr },
+        { L"D2Glide.dll", nullptr },
+        { L"D2Lang.dll", nullptr },
+        { L"D2Launch.dll", nullptr },
+        { L"D2MCPClient.dll", nullptr },
+        { L"D2Multi.dll", nullptr },
+        { L"D2Net.dll", nullptr },
+        { L"D2Sound.dll", nullptr },
+        { L"D2Win.dll", nullptr },
+        { L"Fog.dll", nullptr },
+        { L"Ijl11.dll", nullptr },
+        { L"SmackW32.dll", nullptr },
+        { L"Storm.dll", nullptr } };
+
 D2Offset::D2Offset() :
         D2Offset(D2TEMPLATE_DLL_FILES::D2DLL_INVALID, { }) {
 }
 
-D2Offset::D2Offset(Offsets offsets) :
+/*D2Offset::D2Offset(Offsets offsets) :
         D2Offset(D2TEMPLATE_DLL_FILES::D2DLL_INVALID, offsets) {
-}
+}*/
 
 D2Offset::D2Offset(D2TEMPLATE_DLL_FILES dllFile, Offsets offsets) {
     D2Offset::dllFile = dllFile;
@@ -51,7 +75,10 @@ DWORD D2Offset::getCurrentAddress() {
 
     HMODULE baseAddress = gptDllFiles[nDLL].dwAddress;
     if (!baseAddress) {
-        return 0;
+        D2Offset::loadModules();
+        if (!baseAddress) {
+            return 0;
+        }
     }
 
     int offset = getCurrentOffset();
@@ -67,4 +94,19 @@ DWORD D2Offset::getCurrentAddress() {
     }
 
     return address;
+}
+
+bool D2Offset::loadModules() {
+    for (int i = 0; i < (int) D2TEMPLATE_DLL_FILES::D2DLL_INVALID; i++) {
+        DLLBaseStrc* hDllFile = &gptDllFiles[i];
+
+        HMODULE hModule = GetModuleHandleW(hDllFile->wszName);
+        if (!hModule) {
+            hModule = LoadLibraryW(hDllFile->wszName);
+        }
+
+        hDllFile->dwAddress = hModule;
+    }
+
+    return true;
 }
