+++ src/D2Ptrs.h
@@ -41,16 +41,16 @@
 #define D2FUNC(DLL, NAME, RETURN, CONV, ARGS, OFFSETS) \
     typedef RETURN (CONV * DLL##_##NAME##_t) ARGS; \
     static D2Offset DLL##_##NAME##_FUNC_OFFSET OFFSETS; \
-    static DLL##_##NAME##_t DLL##_##NAME = (DLL##_##NAME##_t)(DLLBASE_##DLL + DLL##_##NAME##_FUNC_OFFSET.getCurrentOffset());
+    static DLL##_##NAME##_t DLL##_##NAME = (DLL##_##NAME##_t)(DLL##_##NAME##_FUNC_OFFSET.getCurrentAddress());
 
 #define D2VAR(DLL, NAME, TYPE, OFFSETS) \
     typedef TYPE DLL##_##NAME##_vt; \
     static D2Offset DLL##_##NAME##_VAR_OFFSET OFFSETS; \
-    static DLL##_##NAME##_vt * DLL##_##NAME = (DLL##_##NAME##_vt *)(DLLBASE_##DLL + DLL##_##NAME##_VAR_OFFSET.getCurrentOffset());
+    static DLL##_##NAME##_vt * DLL##_##NAME = (DLL##_##NAME##_vt *)(DLL##_##NAME##_VAR_OFFSET.getCurrentAddress());
 
 #define D2PTR(DLL, NAME, OFFSETS) \
     static D2Offset DLL##_##NAME##_PTR_OFFSET OFFSETS; \
-    static DWORD NAME = (DLLBASE_##DLL + DLL##_##NAME##_PTR_OFFSET.getCurrentOffset());
+    static DWORD DLL##_##NAME = (DLL##_##NAME##_PTR_OFFSET.getCurrentAddress());
 
 /********************************************************************************
  *                                                                               *
@@ -63,13 +63,13 @@
  *   D2CLIENT.DLL POINTERS                                                       *
  *                                                                               *
  *********************************************************************************/
-D2VAR(D2CLIENT, ScreenSizeX, int, ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0xDBC48, 0xF7034, 0, 0, 0, 0 }));
-D2VAR(D2CLIENT, ScreenSizeY, int, ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0xDBC4C, 0xF7038, 0, 0, 0, 0 }));
-D2VAR(D2CLIENT, PanelOffsetX, int, ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x11B9A0, 0x11D354, 0, 0, 0, 0 }));
-D2VAR(D2CLIENT, PanelOffsetY, int, ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x11B9A4, 0x11D358, 0, 0, 0, 0 }));
-D2VAR(D2CLIENT, InventoryArrangeMode, int, ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x11B99C, 0x11D2B4, 0, 0, 0, 0 }));
+D2VAR(D2CLIENT, ScreenSizeX, int, (D2TEMPLATE_DLL_FILES::D2DLL_D2CLIENT, { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0xDBC48, 0xF7034, 0, 0, 0, 0 }));
+D2VAR(D2CLIENT, ScreenSizeY, int, (D2TEMPLATE_DLL_FILES::D2DLL_D2CLIENT, { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0xDBC4C, 0xF7038, 0, 0, 0, 0 }));
+D2VAR(D2CLIENT, PanelOffsetX, int, (D2TEMPLATE_DLL_FILES::D2DLL_D2CLIENT, { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x11B9A0, 0x11D354, 0, 0, 0, 0 }));
+D2VAR(D2CLIENT, PanelOffsetY, int, (D2TEMPLATE_DLL_FILES::D2DLL_D2CLIENT, { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x11B9A4, 0x11D358, 0, 0, 0, 0 }));
+D2VAR(D2CLIENT, InventoryArrangeMode, int, (D2TEMPLATE_DLL_FILES::D2DLL_D2CLIENT, { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x11B99C, 0x11D2B4, 0, 0, 0, 0 }));
 
-D2VAR(D2CLIENT, VideoOptionCellFileStart, void*, ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0xEA568, 0xE48D0, 0, 0, 0, 0 }));
+D2VAR(D2CLIENT, VideoOptionCellFileStart, void*, (D2TEMPLATE_DLL_FILES::D2DLL_D2CLIENT, { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0xEA568, 0xE48D0, 0, 0, 0, 0 }));
 
 /********************************************************************************
  *                                                                               *
@@ -82,22 +82,22 @@ D2VAR(D2CLIENT, VideoOptionCellFileStart, void*, ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
  *   D2GDI.DLL POINTERS                                                          *
  *                                                                               *
  *********************************************************************************/
-D2VAR(D2GDI, ForegroundRenderWidth, int, ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0xCA9C, 0xCA98, 0, 0, 0, 0 }));
+D2VAR(D2GDI, ForegroundRenderWidth, int, (D2TEMPLATE_DLL_FILES::D2DLL_D2GDI, { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0xCA9C, 0xCA98, 0, 0, 0, 0 }));
 
 /********************************************************************************
 *                                                                               *
 *   D2GFX.DLL POINTERS                                                          *
 *                                                                               *
 *********************************************************************************/
-D2FUNC(D2GFX, GetResolutionMode, int, __stdcall, (), ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -10031, -10012, 0, 0, 0, 0 }));
+D2FUNC(D2GFX, GetResolutionMode, int, __stdcall, (), (D2TEMPLATE_DLL_FILES::D2DLL_D2GFX, { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -10031, -10012, 0, 0, 0, 0 }));
 
 /********************************************************************************
 *                                                                               *
 *   D2WIN.DLL POINTERS                                                          *
 *                                                                               *
 *********************************************************************************/
-D2FUNC(D2WIN, LoadCellFile, CellFile*, __fastcall, (const char* szFile, int Type), ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -10111, -10023, 0, 0, 0, 0 }));
-D2PTR(D2WIN, LoadMpq_I, ({ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x7E60, 0x7E50, 0, 0, 0, 0 }));
+D2FUNC(D2WIN, LoadCellFile, CellFile*, __fastcall, (const char* szFile, int Type), (D2TEMPLATE_DLL_FILES::D2DLL_D2WIN, { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -10111, -10023, 0, 0, 0, 0 }));
+D2PTR(D2WIN, LoadMpq_I, (D2TEMPLATE_DLL_FILES::D2DLL_D2WIN, { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0x7E60, 0x7E50, 0, 0, 0, 0 }));
 
 // end of file -----------------------------------------------------------------
 #endif
