+++ src/DLLmain.cpp
@@ -27,51 +27,17 @@
 #define _D2VARS_H
 
 #include "DLLmain.h"
+
+#include <fstream>
+
 #include "D2Patch.h"
 #include "D2Patches.h"
 
-DLLBaseStrc gptDllFiles[] = {
-        { L"Binkw32.dll", nullptr },
-        { L"BnClient.dll", nullptr },
-        { L"D2Client.dll", nullptr },
-        { L"D2CMP.dll", nullptr },
-        { L"D2Common.dll", nullptr },
-        { L"D2DDraw.dll", nullptr },
-        { L"D2Direct3D.dll", nullptr },
-        { L"D2Game.dll", nullptr },
-        { L"D2Gdi.dll", nullptr },
-        { L"D2Gfx.dll", nullptr },
-        { L"D2Glide.dll", nullptr },
-        { L"D2Lang.dll", nullptr },
-        { L"D2Launch.dll", nullptr },
-        { L"D2MCPClient.dll", nullptr },
-        { L"D2Multi.dll", nullptr },
-        { L"D2Net.dll", nullptr },
-        { L"D2Sound.dll", nullptr },
-        { L"D2Win.dll", nullptr },
-        { L"Fog.dll", nullptr },
-        { L"Ijl11.dll", nullptr },
-        { L"SmackW32.dll", nullptr },
-        { L"Storm.dll", nullptr } };
+
 
 void __fastcall D2TEMPLATE_FatalError(LPCWSTR wszMessage) {
     MessageBoxW(NULL, wszMessage, L"D2Template", MB_OK | MB_ICONERROR);
     TerminateProcess(GetCurrentProcess(), -1);
-}
-
-bool __fastcall D2TEMPLATE_LoadModules() {
-    for (int i = 0; i < (int) D2TEMPLATE_DLL_FILES::D2DLL_INVALID; i++) {
-        DLLBaseStrc* hDllFile = &gptDllFiles[i];
-
-        HMODULE hModule = GetModuleHandleW(hDllFile->wszName);
-        if (!hModule) {
-            hModule = LoadLibraryW(hDllFile->wszName);
-        }
-
-        hDllFile->dwAddress = hModule;
-    }
-
-    return true;
 }
 
 bool __fastcall D2TEMPLATE_GetDebugPrivilege() {
@@ -113,9 +79,15 @@ bool __stdcall DllAttach() {
         return false;
     }
 
-    if (!D2TEMPLATE_LoadModules()) {
+    if (!D2Offset::loadModules()) {
         D2TEMPLATE_FatalError(L"Failed to load modules");
         return false;
+    }
+
+    std::ifstream in(D2HD::Config::DEFAULT_ARCHIVE_NAME);
+    if (in.good()) {
+        in.close();
+        D2HD::Draw::d2mrArchive = loadMPQ(5000, "D2HD.dll", D2HD::Config::DEFAULT_ARCHIVE_NAME.c_str(), "SlashDiabloHD", 0, nullptr);
     }
 
     D2HD::Config::readConfig();
